Agentic_Rag/requirements.txt
anthropic>=0.28.0
chromadb>=0.5.0
tavily-python>=0.3.0
pandas>=2.1.0
rich>=13.7.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
pypdf>=4.0.0
